+++ src/app/pages/calendar/calendar-header.component.ts
@@ -42,7 +42,6 @@ export class CalendarHeaderComponent {
 
     console.log('gotoModeConfig this.entityId: ', JSON.stringify(this.entityFiltered));
 
-
     this.dialogConfig.data = { view: this.view, entity: this.entityFiltered };
     this.dialogConfig.width = '500px';
     this.dialogConfig.height = '500px';